Create SentinelOneMonitorMac
Check for SentinelOne files and processes to determine if installed. Also checks if SentinelOne has Full Disk Access (FDA) which has to be granted by the user through the GUI as a required permission and if the last check in time is over 24 hours. Written by Ricky Cobb, Datto Community.